+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.2.0
+-----
+- Tabularize::Table class for easy tabularization (doh!)
+- Correctly handles ANSI codes
+
 0.1.1
 -----
 - :unicode_display option added for CJK wide characters.